Rename Global.h to global.h
There was a large mismatch on which files included "Global.h" and which "global.h" but in the end the lowercase won it, just because all of the rest of the headers are lowercase.
