+++ src/Windows Code Release 3/Scen Ed/tactions.cpp
@@ -3,7 +3,7 @@
 #include <math.h>
 #include <stdio.h>
 #include <string.h>
-#include "Global.h"
+#include "global.h"
 #include "graphics.h"
 #include "tactions.h"
 #include "edsound.h"
