+++ src/Windows Code Release 3/Scen Ed/townout_stub.cpp
@@ -1,7 +1,7 @@
 #include <Windows.h>
 #include <stdio.h>
 #include <string.h>
-#include "Global.h"
+#include "global.h"
 
 void init_town(short size)
 {
